+++ packages/drift_sync_core/lib/src/sync_dependency_manager.dart
@@ -70,4 +70,3 @@ class DefaultSyncDependencyManager extends SyncDependencyManagerBase {
     _successfullySynced.clear();
   }
 }
-